Package sl1user have the routines for /api/account
